1. move DatePicker component to dna frontend package 2. handle click on outside context menu in List view mode 3. in provider form interchange planningIt and LCO fields 4. remove separate component DataOriginating and include the code blocks within transnational component
