+++ packages/data-product-mfe/src/components/DataProducts.style.scss
@@ -140,6 +140,7 @@
     justify-content: center;
     border: 1px solid #697582;
     padding: 20px;
+    cursor: pointer;
     .addicon {
         width: 40px;
         height: 40px;
