+++ packages/frontend/src/components/DatePicker/DatePicker.scss
@@ -89,6 +89,7 @@
                         }
                     }
                 }
+                .MuiMonthPicker-root .PrivatePickersMonth-root,
                 .PrivatePickersYear-root .PrivatePickersYear-yearButton {
                     &:hover {
                         background-color: #383f49;
